requirements.txt
browser-use>=0.8.1
gradio>=5.49.1
pyyaml>=6.0.2
playwright>=1.55.0
ollama>=0.6.0
